details_screen and favourite_screen
The details_screen.dart file displays detailed information about a selected movie, including its title, genre, release year, description, and poster image. It provides users with a focused view of a movie and may include a button to mark or unmark it as a favorite.

The favourites_screen.dart file shows a list of all movies marked as favorites by the user. It allows users to quickly access and manage their favorite movies, with options to view details or remove them from the favorites list.